fix: naming and code style
