+++ to-do-list/package.json
@@ -33,6 +33,7 @@
     "@reduxjs/toolkit": "^1.5.1",
     "babel-cli": "^6.18.0",
     "babel-preset-react-app": "^3.1.2",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "formik": "^2.2.9",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
