+++ to-do-list/src/components/task-container/TaskContainer.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-export const TaskContainer = (props) => {
+export const TaskContainer = props => {
   const { task } = props;
   return (
     <p>{task.text} {task.status ? "done" : "inProgress"}</p>
