+++ to-do-list/src/db/api.js
@@ -1,0 +1,9 @@
+export const api = (data) => {
+  return {
+    addTask() {
+      return new Promise((resolve) => {
+        setTimeout(() => resolve(data), 4000);
+      })
+    }
+  }
+}
