--- to-do-list/src/db/queries.js
@@ -1,5 +1,0 @@
-export const API = (data) => {
-  return new Promise((resolve) => {
-    setTimeout(() => resolve(data), 4000);
-  })
-}
