--- to-do-list/src/features/validation/validate-task-text.js
@@ -1,7 +1,0 @@
-export const vallidateTaskText = (value) => {
-  let error;
-  if (!value.trim()) {
-    error = "Required";
-  }
-  return error;
-}
