+++ to-do-list/src/features/validation/vallidate.js
@@ -1,0 +1,1 @@
+export const vallidate = value => !value.trim() ? "Required" : "";
