+++ to-do-list/src/index.js
@@ -1,11 +1,12 @@
 import React from "react";
 import { render } from "react-dom";
-import App from "./components/app/App";
+import App from "./components/App/App";
 import store from "./redux/store";
 import { Provider } from "react-redux";
 
 render(
   <Provider store={store}>
-    <App/>
+  <App/>
   </Provider>,
-document.getElementById("root"));
+  document.getElementById("root")
+);
