+++ to-do-list/src/redux/actions/addNewTask.js
@@ -1,0 +1,9 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import { api} from "../../db/api";
+
+export const addNewTask = createAsyncThunk(
+  "taskList/addNewTask",
+  (task) => {
+    return api(task).addTask();
+  }
+);
