+++ to-do-list/src/redux/tasks/taskListSlice.js
@@ -1,0 +1,18 @@
+import { createSlice} from "@reduxjs/toolkit";
+import uuid from "uuid";
+import { defaultTasks } from "../initial-state";
+import { addNewTask } from "./../actions/addNewTask.js";
+ 
+export const taskListSlice = createSlice({
+  name: "taskList",
+  initialState: { tasks: defaultTasks },
+  reducers: {},
+  extraReducers: {
+    [addNewTask.fulfilled]: (state, action) => {
+      action.payload.id = uuid.v4();
+      state.tasks.push(action.payload)
+    },
+  },
+});
+
+export default taskListSlice.reducer;
